Fixed bad filename extraction from path.

+++ source/Path.cpp
@@ -3,8 +3,8 @@
 
 const char* Path::FindFileName(const char* path)
 {
-    size_t length = strlen(path);
+    int length = (int)strlen(path);
     const char* p = path + length - 1;
-    for (size_t i = length - 1; i >= 0 && *p != '/' && *p != '\\'; i--, p--);
+    for (int i = length - 1; i >= 0 && *p != '/' && *p != '\\'; i--, p--);
     return p + 1;
 }
